fix: Resolves an issue where the progress bar for recently watched components is not displayed

+++ composeApp/src/commonMain/kotlin/com/jankinwu/fntv/client/data/model/response/PlayDetailResponse.kt
@@ -55,7 +55,7 @@ data class PlayDetailResponse(
     @param:JsonProperty("ancestor_category")
     val ancestorCategory: String?,
     @param:JsonProperty("ts")
-    val ts: Int,
+    val ts: Long,
     @param:JsonProperty("duration")
     val duration: Int,
     @param:JsonProperty("single_child_guid")
